third_party/blink/web_tests/platform/generic/fast/css/invalidation/reattach-with-sibling-invalidation-expected.txt

A reattach should not discard a scheduled sibling invalidation.

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".

PASS getComputedStyle(document.querySelector('.b')).color is "rgb(255, 0, 0)"
PASS internals.updateStyleAndReturnAffectedElementCount() is 2
PASS getComputedStyle(document.querySelector('.b')).color is "rgb(0, 128, 0)"
PASS successfullyParsed is true

TEST COMPLETE

Should be green
